+++ lib/intl_phone_field.dart
@@ -1,17 +1,25 @@
 library intl_phone_field;
 
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:intl_phone_field/country_picker_dialog.dart';
 
 import './countries.dart';
 import './phone_number.dart';
 
-// ignore: must_be_immutable
 class IntlPhoneField extends StatefulWidget {
+  /// Whether to hide the text being edited (e.g., for passwords).
   final bool obscureText;
+
+  /// How the text should be aligned horizontally.
   final TextAlign textAlign;
+
+  /// How the text should be aligned vertically.
+  final TextAlignVertical? textAlignVertical;
   final VoidCallback? onTap;
-  
+
   /// {@macro flutter.widgets.editableText.readOnly}
   final bool readOnly;
   final FormFieldSetter<PhoneNumber>? onSaved;
@@ -26,8 +34,9 @@ class IntlPhoneField extends StatefulWidget {
   ///    which are more specialized input change notifications.
   final ValueChanged<PhoneNumber>? onChanged;
   final ValueChanged<PhoneNumber>? onCountryChanged;
-  final FormFieldValidator<String>? validator;
-  final bool autoValidate;
+
+  /// For validator to work, turn [autovalidateMode] to [AutovalidateMode.onUserInteraction]
+  final FutureOr<String?> Function(String?)? validator;
 
   /// {@macro flutter.widgets.editableText.keyboardType}
   final TextInputType keyboardType;
@@ -86,8 +95,10 @@ class IntlPhoneField extends StatefulWidget {
   ///    [TextInputAction.previous] for [textInputAction].
   final void Function(String)? onSubmitted;
 
-  /// If false the text field is "disabled": it ignores taps and its
-  /// [decoration] is rendered in grey.
+  /// If false the widget is "disabled": it ignores taps, the [TextFormField]'s
+  /// [decoration] is rendered in grey,
+  /// [decoration]'s [InputDecoration.counterText] is set to `""`,
+  /// and the drop down icon is hidden no matter [showDropdownIcon] value.
   ///
   /// If non-null this property overrides the [decoration]'s
   /// [Decoration.enabled] property.
@@ -97,8 +108,8 @@ class IntlPhoneField extends StatefulWidget {
   ///
   /// This setting is only honored on iOS devices.
   ///
-  /// If unset, defaults to the brightness of [ThemeData.primaryColorBrightness].
-  final Brightness keyboardAppearance;
+  /// If unset, defaults to the brightness of [ThemeData.brightness].
+  final Brightness? keyboardAppearance;
 
   /// Initial Value for the field.
   /// This property can be used to pre-fill the field.
@@ -117,7 +128,7 @@ class IntlPhoneField extends StatefulWidget {
   ///
   /// Specify null to remove the decoration entirely (including the
   /// extra padding introduced by the decoration to save space for the labels).
-  final InputDecoration? decoration;
+  final InputDecoration decoration;
 
   /// The style to use for the text being edited.
   ///
@@ -125,71 +136,137 @@ class IntlPhoneField extends StatefulWidget {
   ///
   /// If null, defaults to the `subtitle1` text style from the current [Theme].
   final TextStyle? style;
+
+  /// Disable view Min/Max Length check
+  final bool disableLengthCheck;
+
+  /// Won't work if [enabled] is set to `false`.
   final bool showDropdownIcon;
 
   final BoxDecoration dropdownDecoration;
 
+  /// The style use for the country dial code.
+  final TextStyle? dropdownTextStyle;
+
   /// {@macro flutter.widgets.editableText.inputFormatters}
   final List<TextInputFormatter>? inputFormatters;
 
-  /// Placeholder Text to Display in Searchbar for searching countries
+  /// The text that describes the search input field.
+  ///
+  /// When the input field is empty and unfocused, the label is displayed on top of the input field (i.e., at the same location on the screen where text may be entered in the input field).
+  /// When the input field receives focus (or if the field is non-empty), the label moves above (i.e., vertically adjacent to) the input field.
   final String searchText;
 
-  /// Color of the country code
-  final Color? countryCodeTextColor;
+  /// Position of an icon [leading, trailing]
+  final IconPosition dropdownIconPosition;
 
-  /// Color of the drop down arrow
-  final Color? dropDownArrowColor;
-
-  /// Icon of the Drop down arrow
-  final IconData? dropDownArrowIcon;
+  /// Icon of the drop down button.
+  ///
+  /// Default is [Icon(Icons.arrow_drop_down)]
+  final Icon dropdownIcon;
 
   /// Whether this text field should focus itself if nothing else is already focused.
   final bool autofocus;
 
-  TextInputAction? textInputAction;
+  /// Autovalidate mode for text form field
+  final AutovalidateMode? autovalidateMode;
 
-  IntlPhoneField(
-      {this.initialCountryCode,
-      this.obscureText = false,
-      this.textAlign = TextAlign.left,
-      this.onTap,
-      this.readOnly = false,
-      this.initialValue,
-      this.keyboardType = TextInputType.number,
-      this.autoValidate = true,
-      this.controller,
-      this.focusNode,
-      this.decoration,
-      this.style,
-      this.onSubmitted,
-      this.validator,
-      this.onChanged,
-      this.countries,
-      this.onCountryChanged,
-      this.onSaved,
-      this.showDropdownIcon = true,
-      this.dropdownDecoration = const BoxDecoration(),
-      this.inputFormatters,
-      this.enabled = true,
-      this.keyboardAppearance = Brightness.light,
-      this.searchText = 'Search by Country Name',
-      this.countryCodeTextColor,
-      this.dropDownArrowColor,
-      this.dropDownArrowIcon,
-      this.autofocus = false,
-      this.textInputAction});
+  /// Whether to show or hide country flag.
+  ///
+  /// Default value is `true`.
+  final bool showCountryFlag;
+
+  /// Message to be displayed on autoValidate error
+  ///
+  /// Default value is `Invalid Mobile Number`.
+  final String? invalidNumberMessage;
+
+  /// The color of the cursor.
+  final Color? cursorColor;
+
+  /// How tall the cursor will be.
+  final double? cursorHeight;
+
+  /// How rounded the corners of the cursor should be.
+  final Radius? cursorRadius;
+
+  /// How thick the cursor will be.
+  final double cursorWidth;
+
+  /// Whether to show cursor.
+  final bool? showCursor;
+
+  /// The padding of the Flags Button.
+  ///
+  /// The amount of insets that are applied to the Flags Button.
+  ///
+  /// If unset, defaults to [EdgeInsets.zero].
+  final EdgeInsetsGeometry flagsButtonPadding;
+
+  /// The type of action button to use for the keyboard.
+  final TextInputAction? textInputAction;
+
+  /// Optional set of styles to allow for customizing the country search
+  /// & pick dialog
+  final PickerDialogStyle? pickerDialogStyle;
+
+  IntlPhoneField({
+    Key? key,
+    this.initialCountryCode,
+    this.obscureText = false,
+    this.textAlign = TextAlign.left,
+    this.textAlignVertical,
+    this.onTap,
+    this.readOnly = false,
+    this.initialValue,
+    this.keyboardType = TextInputType.phone,
+    this.controller,
+    this.focusNode,
+    this.decoration = const InputDecoration(),
+    this.style,
+    this.dropdownTextStyle,
+    this.onSubmitted,
+    this.validator,
+    this.onChanged,
+    this.countries,
+    this.onCountryChanged,
+    this.onSaved,
+    this.showDropdownIcon = true,
+    this.dropdownDecoration = const BoxDecoration(),
+    this.inputFormatters,
+    this.enabled = true,
+    this.keyboardAppearance,
+    @Deprecated('Use searchFieldInputDecoration of PickerDialogStyle instead')
+        this.searchText = 'Search country',
+    this.dropdownIconPosition = IconPosition.leading,
+    this.dropdownIcon = const Icon(Icons.arrow_drop_down),
+    this.autofocus = false,
+    this.textInputAction,
+    this.autovalidateMode = AutovalidateMode.onUserInteraction,
+    this.showCountryFlag = true,
+    this.cursorColor,
+    this.disableLengthCheck = false,
+    this.flagsButtonPadding = EdgeInsets.zero,
+    this.invalidNumberMessage,
+    this.cursorHeight,
+    this.cursorRadius = Radius.zero,
+    this.cursorWidth = 2.0,
+    this.showCursor = true,
+    this.pickerDialogStyle,
+  }) : super(key: key);
 
   @override
   _IntlPhoneFieldState createState() => _IntlPhoneFieldState();
 }
 
 class _IntlPhoneFieldState extends State<IntlPhoneField> {
-  late List<Map<String, dynamic>> _countryList;
-  late Map<String, dynamic> _selectedCountry;
-  late List<Map<String, dynamic>> filteredCountries;
+  late List<Country> _countryList;
+  late Country _selectedCountry;
+  late List<Country> filteredCountries;
+  late String number;
+  bool hasChanged = false;
 
-  FormFieldValidator<String>? validator;
+  String? validationMessage;
 
   @override
   void initState() {
@@ -197,18 +274,30 @@ class _IntlPhoneFieldState extends State<IntlPhoneField> {
     _countryList = widget.countries == null
         ? countries
         : countries
-            .where((country) => widget.countries!.contains(country['code']))
+            .where((country) => widget.countries!.contains(country.code))
             .toList();
     filteredCountries = _countryList;
-    _selectedCountry = _countryList.firstWhere(
-        (item) => item['code'] == (widget.initialCountryCode ?? 'US'),
-        orElse: () => _countryList.first);
-
-    validator = widget.autoValidate
-        ? ((value) => value != null && value.length != 10
-            ? 'Invalid Mobile Number'
-            : null)
-        : widget.validator;
+    number = widget.initialValue ?? '';
+    if (widget.initialCountryCode == null && number.startsWith('+')) {
+      number = number.substring(1);
+      // parse initial value
+      _selectedCountry = countries.firstWhere(
+          (country) => number.startsWith(country.dialCode),
+          orElse: () => _countryList.first);
+      number = number.substring(_selectedCountry.dialCode.length);
+    } else {
+      _selectedCountry = _countryList.firstWhere(
+          (item) => item.code == (widget.initialCountryCode ?? 'US'),
+          orElse: () => _countryList.first);
+    }
+    if (widget.autovalidateMode == AutovalidateMode.always) {
+      var x = widget.validator?.call(widget.initialValue);
+      if (x is String) {
+        setState(() => validationMessage = x);
+      } else {
+        (x as Future).then((msg) => setState(() => validationMessage = msg));
+      }
+    }
   }
 
   Future<void> _changeCountry() async {
@@ -217,131 +306,89 @@ class _IntlPhoneFieldState extends State<IntlPhoneField> {
       context: context,
       useRootNavigator: false,
       builder: (context) => StatefulBuilder(
-        builder: (ctx, setState) => Dialog(
-          child: Container(
-            padding: EdgeInsets.all(10),
-            child: Column(
-              children: <Widget>[
-                TextField(
-                  decoration: InputDecoration(
-                    suffixIcon: Icon(Icons.search),
-                    labelText: widget.searchText,
-                  ),
-                  onChanged: (value) {
-                    setState(() {
-                      filteredCountries = _countryList
-                          .where((country) => country['name']!
-                              .toLowerCase()
-                              .contains(value.toLowerCase()))
-                          .toList();
-                    });
-                  },
-                ),
-                SizedBox(height: 20),
-                Expanded(
-                  child: ListView.builder(
-                    shrinkWrap: true,
-                    itemCount: filteredCountries.length,
-                    itemBuilder: (ctx, index) => Column(
-                      children: <Widget>[
-                        ListTile(
-                          leading: Image.asset(
-                            'assets/flags/${filteredCountries[index]['code']!.toLowerCase()}.png',
-                            package: 'intl_phone_field',
-                            width: 32,
-                          ),
-                          title: Text(
-                            filteredCountries[index]['name']!,
-                            style: TextStyle(fontWeight: FontWeight.w700),
-                          ),
-                          trailing: Text(
-                            '+${filteredCountries[index]['dial_code']}',
-                            style: TextStyle(fontWeight: FontWeight.w700),
-                          ),
-                          onTap: () {
-                            _selectedCountry = filteredCountries[index];
-
-                            if (widget.onCountryChanged != null) {
-                              widget.onCountryChanged!(
-                                PhoneNumber(
-                                  countryISOCode: _selectedCountry['code'],
-                                  countryCode:
-                                      '+${_selectedCountry['dial_code']}',
-                                  number: '',
-                                ),
-                              );
-                            }
-
-                            Navigator.of(context).pop();
-                          },
-                        ),
-                        Divider(thickness: 1),
-                      ],
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
+        builder: (ctx, setState) => CountryPickerDialog(
+          style: widget.pickerDialogStyle,
+          filteredCountries: filteredCountries,
+          searchText: widget.searchText,
+          countryList: _countryList,
+          selectedCountry: _selectedCountry,
+          onCountryChanged: (Country country) {
+            _selectedCountry = country;
+            widget.onCountryChanged?.call(
+              PhoneNumber(
+                countryISOCode: country.code,
+                countryCode: '+${country.dialCode}',
+                number: '',
+              ),
+            );
+            setState(() {});
+          },
         ),
       ),
     );
-    setState(() {});
+    if (this.mounted) setState(() {});
   }
 
   @override
   Widget build(BuildContext context) {
-    return Row(
-      children: <Widget>[
-        _buildFlagsButton(),
-        SizedBox(width: 8),
-        Expanded(
-          child: TextFormField(
-            initialValue: widget.initialValue,
-            readOnly: widget.readOnly,
-            obscureText: widget.obscureText,
-            textAlign: widget.textAlign,
-            onTap: () {
-              if (widget.onTap != null) widget.onTap!();
-            },
-            controller: widget.controller,
-            focusNode: widget.focusNode,
-            onFieldSubmitted: (s) {
-              if (widget.onSubmitted != null) widget.onSubmitted!(s);
-            },
-            decoration: widget.decoration,
-            style: widget.style,
-            onSaved: (value) {
-              if (widget.onSaved != null)
-                widget.onSaved!(
-                  PhoneNumber(
-                    countryISOCode: _selectedCountry['code'],
-                    countryCode: '+${_selectedCountry['dial_code']}',
-                    number: value,
-                  ),
-                );
-            },
-            onChanged: (value) {
-              if (widget.onChanged != null)
-                widget.onChanged!(
-                  PhoneNumber(
-                    countryISOCode: _selectedCountry['code'],
-                    countryCode: '+${_selectedCountry['dial_code']}',
-                    number: value,
-                  ),
-                );
-            },
-            validator: validator,
-            maxLength: _selectedCountry['max_length'],
-            keyboardType: widget.keyboardType,
-            inputFormatters: widget.inputFormatters,
-            enabled: widget.enabled,
-            keyboardAppearance: widget.keyboardAppearance,
-            autofocus: widget.autofocus,
-            textInputAction: widget.textInputAction,
+    return TextFormField(
+      initialValue: number,
+      readOnly: widget.readOnly,
+      obscureText: widget.obscureText,
+      textAlign: widget.textAlign,
+      textAlignVertical: widget.textAlignVertical,
+      cursorColor: widget.cursorColor,
+      onTap: widget.onTap,
+      controller: widget.controller,
+      focusNode: widget.focusNode,
+      cursorHeight: widget.cursorHeight,
+      cursorRadius: widget.cursorRadius,
+      cursorWidth: widget.cursorWidth,
+      showCursor: widget.showCursor,
+      onFieldSubmitted: widget.onSubmitted,
+      decoration: widget.decoration.copyWith(
+        prefixIcon: _buildFlagsButton(),
+        counterText: !widget.enabled ? '' : null,
+      ),
+      style: widget.style,
+      onSaved: (value) {
+        widget.onSaved?.call(
+          PhoneNumber(
+            countryISOCode: _selectedCountry.code,
+            countryCode: '+${_selectedCountry.dialCode}',
+            number: value!,
           ),
-        ),
-      ],
+        );
+      },
+      onChanged: (value) async {
+        hasChanged = true;
+        final phoneNumber = PhoneNumber(
+          countryISOCode: _selectedCountry.code,
+          countryCode: '+${_selectedCountry.dialCode}',
+          number: value,
+        );
+        // validate here to take care of async validation
+        var msg;
+        if (widget.autovalidateMode != AutovalidateMode.disabled) {
+          msg = widget.disableLengthCheck ||
+                  value.length >= _selectedCountry.minLength &&
+                      value.length <= _selectedCountry.maxLength
+              ? null
+              : (widget.invalidNumberMessage ?? 'Invalid Mobile Number');
+          msg ??= await widget.validator?.call(phoneNumber.completeNumber);
+          setState(() => validationMessage = msg);
+        }
+        widget.onChanged?.call(phoneNumber);
+      },
+      validator: (value) => validationMessage,
+      maxLength: widget.disableLengthCheck ? null : _selectedCountry.maxLength,
+      keyboardType: widget.keyboardType,
+      inputFormatters: widget.inputFormatters,
+      enabled: widget.enabled,
+      keyboardAppearance: widget.keyboardAppearance,
+      autofocus: widget.autofocus,
+      textInputAction: widget.textInputAction,
+      autovalidateMode: widget.autovalidateMode,
     );
   }
 
@@ -351,37 +398,48 @@ class _IntlPhoneFieldState extends State<IntlPhoneField> {
       child: InkWell(
         borderRadius: widget.dropdownDecoration.borderRadius as BorderRadius?,
         child: Padding(
-          padding: const EdgeInsets.symmetric(vertical: 8),
+          padding: widget.flagsButtonPadding,
           child: Row(
+            mainAxisSize: MainAxisSize.min,
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              if (widget.showDropdownIcon) ...[
-                Icon(
-                  widget.dropDownArrowIcon ?? Icons.arrow_drop_down,
-                  color: widget.dropDownArrowColor,
-                ),
-                SizedBox(width: 4)
+              if (widget.enabled &&
+                  widget.showDropdownIcon &&
+                  widget.dropdownIconPosition == IconPosition.leading) ...[
+                widget.dropdownIcon,
+                SizedBox(width: 4),
               ],
-              Image.asset(
-                'assets/flags/${_selectedCountry['code']!.toLowerCase()}.png',
-                package: 'intl_phone_field',
-                width: 32,
-              ),
-              SizedBox(width: 8),
+              if (widget.showCountryFlag) ...[
+                Image.asset(
+                  'assets/flags/${_selectedCountry.code.toLowerCase()}.png',
+                  package: 'intl_phone_field',
+                  width: 32,
+                ),
+                SizedBox(width: 8),
+              ],
               FittedBox(
                 child: Text(
-                  '+${_selectedCountry['dial_code']}',
-                  style: TextStyle(
-                      fontWeight: FontWeight.w700,
-                      color: widget.countryCodeTextColor),
+                  '+${_selectedCountry.dialCode}',
+                  style: widget.dropdownTextStyle,
                 ),
               ),
+              if (widget.enabled &&
+                  widget.showDropdownIcon &&
+                  widget.dropdownIconPosition == IconPosition.trailing) ...[
+                SizedBox(width: 4),
+                widget.dropdownIcon,
+              ],
               SizedBox(width: 8),
             ],
           ),
         ),
-        onTap: _changeCountry,
+        onTap: widget.enabled ? _changeCountry : null,
       ),
     );
   }
+}
+
+enum IconPosition {
+  leading,
+  trailing,
 }